[llvm] Make taichi's Ndarray carry a ptr to its DeviceAllocation.
Sorry I left a few todos in this PR since it enables some unification
between cpu/cuda backend, as well as some cleanups.
But they're not relevant to the goal of this PR so I'll send them
separately.

This PR mainly enforces an invariant that:
- Taichi ndarrays' ptr points to a DeviceAllocation instead of the raw
data like torch-based ndarray.

The invariant above makes it easier to unify our Ndarray implementation
among multiple backends. I'll add an opengl implementation of ndarray
on top of this.

note: we keep the data_ptr_ pointing to a raw ptr (so added a new ptr
for DeviceAllocation* instead) to avoid breaking TNG.

[ghstack-poisoned]