[ci] Run cpp tests in unix_docker_test.sh.
The current situation is a bit confusing that we run c++ tests in
unix_docker_build.sh. It didn't run in unix_docker_test.sh since we
didn't copy the build/ into docker image. This PR fixes the issue that
we now run c++ tests in unix_docker_test.sh.