[opengl] Create new args buf instead of reusing one args buf.
When reusing one args_buf for different gl kernels we spend a lot of
time on glMapBufferRange() in syncing. But these syncs are indeed not
necessary. Now we just create one args_buf per kernel launch.
This improve mpm88 benchmark from 16.67fps to 30fps.

Thanks to @k-ye and @bobcao3 for helping!

ghstack-source-id: 7a2e2f0051dc52339d1fabac7f82148b4416814d
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/3835