[opengl] Allocate new arg bufs per kernel launch
When reusing one args_buf for different gl kernels we spend a lot of
time on glMapBufferRange() in syncing. But these syncs are indeed not
necessary. Now we just create one args_buf per kernel launch.
This improve mpm88 benchmark from 16.67fps to 33fps.

Thanks to @k-ye and @bobcao3 for helping!