[opengl] Do not promote simple ExternalTensorShapeAlongAxisStmt into globals.
Previously we only make sure `for i in range(x)` (here `x` is ndarray)
don't get their end_range from gtmp on opengl. This PR expands the scope
to all simple calculations involving only ConstStmt and
ExternalTensorShapeAlongAxisStmt as leaves, and it's not limited to the
end_stmt in RangeForStmt as well. In most cases this should save an
additional serial offloaded task which only saves the simple
intermediates into gtmp.

[ghstack-poisoned]