[docs] add the step of setting "TI_WITH_VULKAN" for linux
It seems currently when building from source, by default, taichi will not enable Vulkan backend. Only when the cmake flag "TI_WITH_VULKAN" set to `on` will it be enabled. See  https://github.com/taichi-dev/taichi/blob/00183bbe9b966727c2a28fdfadb17286ba90eb17/cmake/TaichiCore.cmake#L8 . 
As a result, I was unable to use Vulkan backend for my custom build of taichi, and finding out this took me a long time. For some strange reasons, the step of setting this cmake flag is shown in the Windows section, but not in the Linux section. This commit adds the missing step of setting cmake flags back to the Linux section