[lang] Annotate constants with dtype without casting.
Fixes #4210

This PR also simplifies the clamping logic by:
1. Make sure constant is valid in range of specified dtype.
2. Always use int64/float64 to communicate between python and C++ and
converts it to the specified dtype in C++. This is just to avoid
handling each dtype on both python and C++ side.

We might consider further improve it by moving this logic from python to C++ in the future and this PR made it a step closer.