[lang] Make ti.cfg an alias of runtime cfg (#4264)
Before this PR we have `ti.cfg` and `ti.lang.impl.current_cfg()` and
they're different objects. It's confusing to have both so let's just
make `ti.cfg` an alias of the other.

Note `ti.cfg` used to be default_config before `ti.init()` is called and
now it's `None`. It's safe to chagne this behavior as nothing can be
done without `ti.init()`.