[bug] Update children_offsets & stride info to align as elem_stride
Currently we run 2 passes to generate snode tree layout in opengl
codegen.
1. One pass from leaf to root so that we can calculate the real size &
elem_stride for each snode and sort the children from small to large.
2. Another pass from root to leaf so that we can align as elem_stride.

Note we cannot eliminate the first pass due to
https://github.com/taichi-dev/taichi/blob/master/taichi/backends/opengl/struct_opengl.cpp#L50.

The fix for the alignment issue was simple: we used children_offsets
information in the codegen but didn't update it in the second pass.
This PR fixes the issue and also reorganizes the code a bit to make it
easier to read and understand.

ghstack-source-id: 958b5ed2293af69ab1e2a8e7f734a651fab37f36
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/4345