[build] Default symbol visibility to hidden for all targets
Taichi can be used by external applications that has conflicting
dependencies such as ImGui or Vulkan. In order to avoid any conflicts or
nasty bugs, enforce the visibility to 'hidden' by default so only the
public API is exported.

This moves the config to the global property instead of the per target
property