Refactored README.md
The current readme is packed a bit more details than is required (for example IDE for Windows) but also several references shy of completion. A README should also serve as a hub of information for the community. In that sense, I updated this document to improve scannability and reduce redundancy that the extra details entail. 

If I were given more time, I would consider building a *separate* GitHub repo for hosting these miscellaneous demos. For now we have way too many and, in this doc, we only *cobbled together* these repos. Offtopic: please bear with me.. We have scores of sub-repos under taichi-dev. We could consider sorting out the relationship of these repos. For example, we have a community repo, but it is rarely used. Is this repo international?