[aot] Support return in vulkan aot
Vulkan aot return didn't work since it was trying to serialize
`Datatype` which was a raw pointer. This PR fixes the issue by changing
it to a `PrimitiveTypeID` which is a `int`. Note it's possible that we
limit general `Kernel` args and rets to be primitive types only but leaving that
to a followup PR to avoid making this PR too complicated.