[dx11] Fix device_ lifetime and add an HLSL flag
1. Change Dx11Pipeline::device_ from a shared_ptr to a raw pointer
Previously, Dx11Pipeline::device_ was a shared_ptr.
This can cause the device_ to be destructed twice when the program
finishes.
Because the pipeline is created by Dx11Device::create_pipeline,
logically speaking the device owns the pipeline, rather than the
pipeline owning the device. So it makes sense if the pipeline is not
tracking the life time of its Dx11Device object, thus switching to a
raw pointer.

2. Call hlsl.remap_num_workgroups_builtin()
This fixes the "RuntimeError: NumWorkgroups builtin is used, but
remap_num_workgroups_builtin() was not called. Cannot emit code
for this builtin." error.

3. This change also adds a few other Dx11CommandList commands.