[vulkan] Set kApiVersion to VK_API_VERSION_1_3
This change fixes issue #4900. On certain AMD GPUs (such as RX 5500),
the Vullkan version needs to be 1.3 in order for Volk to correctly populate
the "vkGetDeviceBufferMemoryRequirements" function pointer. If the
version is 1.2, the pointer will be NULL and cause an assertion failure.

Tested by running the voxel challenge program and see the program
working as expected on a Windows PC.