[aot] Bind graph APIs to python and add mpm88 example
This PR supports graph builder and runner APIs in python. Note for
simplicity I've merged builder and runner in the same Python class.
Please feel free to comment if you have any suggestions.

This PR also adds a test of saving mpm88 graph in aot module, as well
as an example script to demonstrate the speed improvement (15fps ->
45fps) compared to the current taichi.

ghstack-source-id: 13059ffa77fe3e0da323804db1cbe86b340e3b06
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/5023