[aot] Support specifying spirv_version in CompileConfig
Ideally we should specify this in `ti.aot.Module` instead of `ti.init`
but tbh we need more refactor before we can compile to a target other
than the current running one. So let's use this to avoid complicating
other parts of the codebase, e.g. offline cache uses CompileConfig as
part of cache key.