[spirv] [bug] Fix invalid CFG error when simulated atomic is present
Fixes #5555

The root cause for the issue above is the input spirv code is invalid so
it errors out when running condtional constant propagation pass in
spirv_opt.

Running `spirv-val` with the generated spirv gives the following error:
```
line:131: OpPhi's incoming basic block <id> 39[%39] is not a predecessor
of <id> 81[%81].
  %tmp11_i32 = OpPhi %int %tmp4_i32 %39 %90 %83
```

The bug happens when a simulated atomic ops is followed by a range-for loop where `curr_label_` isn't set properly.
So when we query `spirv::Label init_label = ir_->current_label();` in
the `RangeForStmt` codegen we get the wrong label to use as incoming
block.

It doesn't error out in vulkan backend since we enable float atomic
extension in vulkan backend when it's available by default so that
simulated atomics codepath was never exercised there.