[vulkan] Relax number of array args for each kernel
fixes #5684

The root cause in #5684 is that we used to have a limit of 64 arguments
in total but only the first 16 can be array args. It turns out it's not
enough for some use cases so let's relax that to 32.

Also just noticed that we used to limit number of args to 8 on opengl
backend. Now that we've move it from ssbo to uniform storage on spirv codegen it
shouldn't be an issue any more.