[opengl] Fix target device for opengl aot
Note we used to generate invalid spirv shaders during opengl aot and
failed to run spirv-opt passes. But it only generated a warning instead
of an error so we didn't catch this in tests.

This was the second bug triggered by invalid spirv shaders. We
probably should just throw when spirv-opt fails. Alternatively we can
run spirv-val separately if we really want to ignore spirv-opt failures.

Note we don't properly handle gles target device yet, will do that in a
followup PR.

ghstack-source-id: be73342c108aa3df3d371835aced5abc010c6860
Pull Request resolved: https://github.com/taichi-dev/taichi/pull/5773