[llvm] [refactor] Split LLVMCompiledData of kernels and tasks (#6019)
Related issue = #5511 
The compiler first compiles every offloaded task to a `LLVMCompiledTask`
which contains an LLVM module, the name of the offloaded task function
inside the module, and some extra information for linking. Then, The
compiler links the modules in the `LLVMCompiledTask`s, the runtime
modules and the struct modules used in the kernel together, and creates
a `LLVMCompiledKernel` that contain the linked LLVM module and the names
of the offloaded tasks.

Both `LLVMCompiledTask` and `LLVMCompiledKernel` need to store the
generated LLVM module and the names of the functions of the offloaded
tasks inside the module. `LLVMCompiledTask` also stores additional
information which will be used when linking like which SNodeTrees are
used and the sizes of TLS buffers used in parallel struct for.
<!--
Thank you for your contribution!

If it is your first time contributing to Taichi, please read our
Contributor Guidelines:
  https://docs.taichi-lang.org/docs/contributor_guide

- Please always prepend your PR title with tags such as [CUDA], [Lang],
[Doc], [Example]. For a complete list of valid PR tags, please check out
https://github.com/taichi-dev/taichi/blob/master/misc/prtags.json.
- Use upper-case tags (e.g., [Metal]) for PRs that change public APIs.
Otherwise, please use lower-case tags (e.g., [metal]).
- More details:
https://docs.taichi-lang.org/docs/contributor_guide#pr-title-format-and-tags

- Please fill in the issue number that this PR relates to.
- If your PR fixes the issue **completely**, use the `close` or `fixes`
prefix so that GitHub automatically closes the issue when the PR is
merged. For example,
    Related issue = close #2345
- If the PR does not belong to any existing issue, free to leave it
blank.
-->

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>