[aot] Support multi-target builds for Apple M1 (#6083)
This PR enables Taichi to be built into multi-target libraries. Recent
releases of macOS supports dynamically linking against x86_64 or arm64
of the same `.dylib` on demand, depending on the arch of the parent
process. Previously this was not possible because the CMake script
forced a `-march=nehalem` that is not a kind of arm64; and
`CMAKE_OSX_ARCHITECTURES` is forced `arm64`.

The multi-target feature is only enabled with
`-DCMAKE_OSX_ARCHITECTURES="arm64;x86_64"`. Python builds will select
one of the archs, based on the arch of the skbuild Python process.

<img width="1260" alt="图片"
src="https://user-images.githubusercontent.com/1487605/190690041-f4a04941-a0dc-440e-b083-c59ae00b111f.png">

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>