[opengl] Fix GLES support (#6121)
Issue: #5769 

### Brief Summary
GLES support was broken since the migration to spirv-based codegen and
this PR is the first step to bring it back. Also added a bunch of
`check_opengl_error` calls to make debugging easier.
Note thorough testing for GLES backend is currently missing and should
be added in followup PRs.