[LLVM] Add runtime overflow detection on LLVM-based backends (#6166)
Issue: #5932 

### Brief Summary
This PR adds runtime overflow detection for addition, subtraction,
multiplication and shift left operations on integer types
(i/u8/16/32/64). This only works when the compiler is GCC or clang
because I've used some built-in functions. This feature is not supported
when compiled with MSVC.
The output of the example in #5932 when `debug=True`:
```
[Taichi] version 1.1.4, llvm 10.0.0, commit f892b138, linux, python 3.9.12
[Taichi] Starting on arch=x64
Multiplication overflow detected in File "/home/lin/test/ndrange.py", line 8, in populate:
    for i, j in ti.ndrange(n, n):
                ^^^^^^^^^^^^^^^^
```
There isn't a built-in overflow detection function to `shl`, so I
implemented it by checking whether `a << b >> b == a`.
The multiplication of power of two is optimized to `shl` in `alg_simp`,
so sometimes the multiplication will be warned shift left overflow, but
I think it should be ok?

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>