[aot] Device capability refactorization (#6184)
This PR refactorized `DeviceCapability` control logics in Taichi.
`DeviceCapability` controls the compilation of kernels and the execution
behavior on device. The current implementation collect capabilities from
the device interface but it blocks extension of AOT code generation
because it depends on the currently attached compute device whether to
enable a set of feature during compilation. AOT users, however, demands
an interface to control the availability of `DeviceCapability`.

This PR does several things and it seems large:
- Removed runtime capability form the list of caps, see code comments
for details of design decisions;
- Encapsulate the capability settings in `Device` in
`DeviceCapabilityConfig`;
- Replace any reference to `Device` with `DeviceCapabilityConfig` in
codegens;
- Removed `TargetDevice` for in the SPIR-V codegen;
- Adapt backends (basically gfx backends) to this change.

Python and C-API users will be able to configure `DeviceCapability`s in
future PRs.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>