[example] Fix out-of-bound access in examples (#6500)
### Brief Summary

Before #6485 out-of-bound accesses of SNodes were silent - through
`BitExtractStmt` they became in-bound but might not result in expected
results. The optimization in #6485 removed the in-bound transition in
certain cases so it was likely that an out-of-bound access would behave
differently. The CI failure of #6485
(https://github.com/taichi-dev/taichi/actions/runs/3367141462/jobs/5584351541)
revealed out-of-bound access in examples, which could be confirmed by
turning on `ti.init(debug=True)`. This PR aims at fixing those examples.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>