[aot] Deprecate filename arg in ti.aot.Module.save() (#6554)
Issue: #

### Brief Summary
`filename` argument in `ti.aot.Module.save()` has been no-op for a long
time. This PR deprecates it and annouce that it'll be removed in v1.4.0
release.

Also added a test file called `test_deprecation.py` which is designed to
be the place to check all DeprecationWarnings so that before every
release we can just check here to see if we can safely remove any
deprecated apis.