[ir] Add ir_traits namespace to use less dynamic casts & Run CFG only ever once (#6812)
This speeds up CFG a little bit by removing most of the dynamic casts,
and create `Load` and `Store` IR Traits that can be applied to
statements that have this trait. (4~5s out of ~1.2 minutes)

In addition, since most of the time in the load/store source/destination
analysis we only return a single statement or even none, using vector as
a return type seems excessive. Here we create a `one_or_more` template
that is a variant between one or more (vector of ones), and we apply the
`begin`, `end`, `empty`, and `size` functions so that this can directly
plug into our code, and we can directly use range-for loops on them.

Then we should disable to `while (!modified)` loop for CFG passes and
only ever run them once in `full_simplify`. I observed no performance
degradation with this change while reducing compile time further down to
around 45s

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>