[Lang] Support  LU sparse solver on CUDA backend (#6967)
Issue: #2906 

### Brief Summary
To be consistent with API on CPU backend, this pr provides LU sparse
solver on CUDA backend. CuSolver just provides a CPU version API of LU
sparse solver which is used in this PR. The cuSolverRF provides a GPU
version LU solve, but it only supports `double` datatype. Thus, it's not
used in this PR.

Besides, the `print_triplets` is refactored to resolve the ndarray
`read` constraints (the `read` and `write` data should be the same
datatype).

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>