Update base for Update on "[ir] Replace FuncCallExpression with FrontendFuncCallStmt"
FuncCall is not just an expression. It executes code and returns the result. It should be more like FrontendAllocaStmt rather than being FuncCallExpression. If it is an expression, the IR printer handles it like an expression and prints the detail of the func call everywhere its return value is used. If it is a statement, it will only be printed once, and other statements/expressions reference its return value by its statement number.

[ghstack-poisoned]