[opt] Automatically identify allocas to scalarize (#7055)
Issue: #2590

### Brief Summary

Previously `ScalarizePointers` (which is for allocas) takes place only
when `dynamic_index=False`. In this PR, we automatically identify those
allocas only indexed with constants and get them scalarized, and let the
remaining allocas go through the code path for dynamic indexing. In this
way, we make old user programs behave the same regardless of the
`dynamic_index` option, and provide dynamic indexing support for new
user programs at the same time.

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lin Jiang <linjiang@taichi.graphics>