Update on "[ir] Let real function return nested StructType"
issue: #6983

Before this PR, if we want to return a struct type in a real function, we flatten the struct in LLVM. For example, `struct(a: Ta, s: struct(b: Tb, c: Tc))` is flattened to `struct(a: Ta, b: Tb, c: Tc)`. After this PR, we do not flatten anymore. We directly return `struct(a: Ta, s: struct(b: Tb, c: Tc))`. 

Note that we still need  to pass the flattened return values from Python to CHI-IR because we can't pass a Python struct into C++. We compose the target struct type from the flattened values and the return type in the codegen.

[ghstack-poisoned]