[metal] Disable a kernel test in offline cache to unblock CI (#7154)
Issue: #6221 

### Brief Summary

`kernel4` in `test_offline_cache.py` triggers a weird LLVM internal
error on metal backend for macos12+. After moving metal runtime to
gfxruntime by @PENGUINLIONG and #7201, this is now the only blocker for
us to run full CI on macos ventura. As discussed with @PENGUINLIONG and
@feisuzhu offline, let's disable this test to unblock CI and add it back
once the issue is fixed.