[opengl] Fix GLFW global context issue (#7230)
Related: #7218

Without this GGUI will mess with the OpenGL backend because both tries
to initialize & destroy GLFW.

There is an issue tho with dependency. Which target should
`rhi/window_system` belong to? Maybe we need a device API target that
collects all the specific backends and generate a static lib with all
backends and optional dependencies like GLFW? (Anyways, that will be a
TODO for another PR)

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>