[Bug] Fix caching same loop invariant global vars inside nested fors (#7285)
Issue: fixes #7263

### Brief Summary
The bug is caused by caching a same global variable in two different
local vars in two nested loops. To avoid this, for a cache-able global
var, we find the outermost loop that it can be moved out of instead of
just moving out of the innermost loop.