Update base for Update on "[ir] Refactor result fetching"
This PR adds code that fetches the return values out of a StructType instead of uint64 arrays. The next PR will let the kernel return a StructType instead of a uint64 array.

Note that before the refactoring, f16 was converted to f32 before returning. However after the refactor, we convert the f16 to f32 on host after it is returned (in the next PR). Therefore, I added an implementation of f16-to-f32 conversion (from stackoverflow) to util/bits.cpp.




[ghstack-poisoned]