Update base for Update on "[llvm] Let the RuntimeContext use the host result buffer"
In this PR, we allocate and deallocate the result buffer on the device when launching instead of using the preallocated memory. However, the preallocated result buffer is still used by the runtime JIT calls. We use an async memory pool on CUDA to allocate and deallocate the result buffer, which is almost as fast as using the preallocated memory.

Async memory operations on AMDGPU is not very stable so we use sync Malloc and Free on AMDGPU.

In following PRs the host result buffer will be managed by `LaunchContextBuilder` instead of `Program`.




[ghstack-poisoned]