[vulkan] Support printing constant strings containing % (#7499)
Issue: #

### Brief Summary
This PR fixes an issue that crashes the vulkan backend when printing
constant strings containing '%'.
It escapes all '%' in the constant strings to "%%", allowing them to be
printed correctly in Vulkan's debug_printf function, since these strings
and other formatting specifiers are concatenated to directly form the
first parameter of the function. In other words, if a wild '%d' or '%s'
is passed to printf, the function will incorrectly assume that there is
a new argument to take (from the stack?), which will mess everything up
and may result in type confusion and buffer overflows.