Handle 2 digit minor versions correctly
The docs previously stated to use `python3 -c "print(__import__('sys').version[:3])"` to verify the Python version you were using. This only works if the minor version is 1 digit (like 3.6), but not for two digits (like 3.10 or 3.11). I've modified it to do this correctly regardless of how many digits the minor version is. 

Incidentally, should this page also be updated to mention 3.11 is supported? I don't have a machine with Python 3.11 installed to test.